+++ src/__test__/core/query/ImmutableQuerySpec.ts
@@ -203,6 +203,18 @@ describe("ImmutableQuery", ()=> {
     expect(query.getFrom()).toEqual(10)
   })
 
+  it("getPage()", ()=> {
+    let query = this.query.setSize(15)
+    expect(query.getPage()).toEqual(1)
+    query = query.setSize(20)
+    expect(query.getPage()).toEqual(1)
+    query = query.setFrom(60)
+    expect(query.getPage()).toEqual(4)
+    // Page should always be an integer
+    query = query.setFrom(50)
+    expect(query.getPage()).toEqual(3) // floor(3.5)
+  })
+
   it("setHighlight()", ()=> {
     let query = this.query.setHighlight({
       "fields": {