+++ src/components/search/filters/input-filter/InputFilter.unit.tsx
@@ -13,11 +13,14 @@ import {
 
 import * as sinon from "sinon";
 
+const omit = require("lodash/omit")
+
 describe("InputFilter tests", () => {
 
   beforeEach(() => {
 
     this.searchkit = SearchkitManager.mock()
+    spyOn(this.searchkit, "performSearch")
 
     this.searchkit.translateFunction = (key)=> {
       return {
@@ -148,9 +151,11 @@ describe("InputFilter tests", () => {
     this.createWrapper(false)
     this.typeSearch('m')
     this.typeSearch('ma')
-    expect(this.accessor.state.getValue()).toBe("ma")
+    // State left in the component
+    expect(this.accessor.state.getValue()).toBe(null)
     expect(spy.callCount).toBe(0)
     this.wrapper.find("form").simulate("submit")
+    expect(this.accessor.state.getValue()).toBe("ma")
     expect(spy.callCount).toBe(1)
   })
 
@@ -185,7 +190,8 @@ describe("InputFilter tests", () => {
       prefixQueryFields:null,
       queryOptions: {},
       prefixQueryOptions: {},
-      queryBuilder:undefined
+      queryBuilder:undefined,
+      onQueryStateChange:jasmine.any(Function)
     })
 
   })
@@ -198,11 +204,12 @@ describe("InputFilter tests", () => {
     expect(options).toEqual({
       title: "Test title",
       addToFilters: true,
-      queryFields: ["title"],
       prefixQueryFields:null,
+      queryFields: ["title"],
       queryOptions: {},
       prefixQueryOptions: {},
-      queryBuilder:undefined
+      queryBuilder:undefined,
+      onQueryStateChange:jasmine.any(Function)
     })
 
   })
@@ -223,7 +230,8 @@ describe("InputFilter tests", () => {
       prefixQueryFields:["prefix"],
       queryOptions:{minimum_should_match:"60%"},
       prefixQueryOptions:{minimum_should_match:"70%"},
-      queryBuilder:QueryString
+      queryBuilder:QueryString,
+      onQueryStateChange:jasmine.any(Function)
     })
   })
 
@@ -232,6 +240,54 @@ describe("InputFilter tests", () => {
       containerComponent: <Panel collapsable={true} />
     })
     expect(hasClass(this.wrapper.find(".sk-panel__header"), "is-collapsable")).toBe(true)
+  })
+
+  describe("url change + blurAction", ()=> {
+
+    it("blurAction:restore", ()=> {
+      this.createWrapper(false, ["title"], ["prefix"], {
+        blurAction:"restore"
+      })
+      this.typeSearch("la")
+      expect(this.wrapper.node.getValue() ).toEqual("la")
+      this.accessor.fromQueryObject({
+        test_id:"foo"
+      })
+      expect(this.wrapper.node.getValue() ).toEqual("foo")
+
+      this.typeSearch("bar")
+      expect(this.wrapper.node.getValue()).toEqual("bar")
+      this.wrapper.find(".sk-input-filter__text")
+        .simulate("blur")
+
+      // should be restored to previous value
+      expect(this.wrapper.node.getValue()).toEqual("foo")
+      expect(this.searchkit.performSearch).not.toHaveBeenCalled()
+
+    })
+
+    it("blurAction:search", ()=> {
+      this.createWrapper(false, ["title"], ["prefix"], {
+        blurAction:"search"
+      })
+      this.typeSearch("la")
+      expect(this.wrapper.node.getValue() ).toEqual("la")
+      this.accessor.fromQueryObject({
+        test_id:"foo"
+      })
+      expect(this.wrapper.node.getValue() ).toEqual("foo")
+
+      this.typeSearch("bar")
+      expect(this.wrapper.node.getValue()).toEqual("bar")
+      this.wrapper.find(".sk-input-filter__text")
+        .simulate("blur")
+
+      // should flush value + search
+      expect(this.wrapper.node.getValue()).toEqual("bar")
+      expect(this.searchkit.performSearch).toHaveBeenCalled()
+
+    })
+
   })